Changement de type de la colonne connectType dans la table users

+++ api/src/database/init.sql
@@ -31,7 +31,7 @@ CREATE TABLE IF NOT EXISTS users (
   password VARCHAR(255) NOT NULL,
   role VARCHAR(255) NOT NULL DEFAULT 'user',
   currency VARCHAR(255) NOT NULL DEFAULT 'EUR',
-  connectType INT NOT NULL DEFAULT 0 -- 0 = local, 1 = discord
+  connectType SMALLINT NOT NULL DEFAULT 0 -- 0 = local, 1 = discord
 );
 
 -- Create table cryptos_users
